+++ setup/update-to-0.6.0.sql
@@ -4,3 +4,7 @@ UPDATE `tasks` SET `priority`='100' WHERE `priority`='low';
 
 ALTER TABLE `tasks`
   MODIFY COLUMN `priority` INT(3) UNSIGNED;
+
+ALTER TABLE `tasks`
+  CHANGE COLUMN `marked` `isDone` tinyint(1) NOT NULL DEFAULT '0';
+