+++ src/Dingbat/Action/Card/CreateTask.php
@@ -29,13 +29,13 @@ class CreateTask extends Action\AbstractImpl
         // create task
         $model           = new Task();
         $model->name     = $request->getParsedBodyParam('name');
-        $model->marked   = (bool) $request->getParsedBodyParam('marked', false);
+        $model->isDone   = (bool) $request->getParsedBodyParam('isDone', false);
         $model->priority = DatabaseUtils::parseTaskPriority($request->getParsedBodyParam('priority', 500));
         $model->cardId   = $card->id;
 
         // validation
         Task::validators()['name']->assert($model->name);
-        Task::validators()['marked']->assert($model->marked);
+        Task::validators()['isDone']->assert($model->isDone);
         Task::validators()['priority']->assert($model->priority);
 
         // save