+++ src/Dingbat/Model/Task.php
@@ -7,11 +7,11 @@ use Respect\Validation\Validator;
 
 
 /**
- * @property int     id
- * @property string  name
- * @property boolean marked
- * @property string  priority
- * @property int     cardId
+ * @property int     $id
+ * @property string  $name
+ * @property boolean $isDone
+ * @property string  $priority
+ * @property int     $cardId
  */
 class Task extends Model
 {
@@ -22,7 +22,7 @@ class Task extends Model
     public static function validators() {
         return [
             'name'     => Validator::stringType()->notEmpty()->setName('name'),
-            'marked'   => Validator::boolType()->setName('marked'),
+            'isDone'   => Validator::boolType()->setName('isDone'),
             'priority' => Validator::intType()->between(1, 999)->setName('priority'),
             'cardId'   => Validator::intType()->notEmpty()->callback(function($v) {
                 return Card::query()->find($v) instanceof Card;