+++ tests/functional/TaskCreateCept.php
@@ -9,7 +9,7 @@ $guy = new TestGuy($scenario);
 $guy->wantTo('create valid task');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 'normal'
 ]);
 $guy->seeResponseIsJson();
@@ -19,15 +19,15 @@ $guy->seeResponseContainsJson(['id' => 4]);
 $guy->sendGET('/tasks/4');
 $guy->seeResponseContainsJson([
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 500
 ]);
 
 
-$guy->wantTo('marked is `true`');
+$guy->wantTo('isDone is `true`');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
-    'marked'   => true,
+    'isDone'   => true,
     'priority' => 'normal'
 ]);
 $guy->seeResponseIsJson();
@@ -36,12 +36,12 @@ $guy->seeResponseContainsJson(['id' => 5]);
 $guy->sendGET('/tasks/5');
 $guy->seeResponseContainsJson([
     'name'     => 'something',
-    'marked'   => true,
+    'isDone'   => true,
     'priority' => 500
 ]);
 
 
-$guy->wantTo('marked is not set');
+$guy->wantTo('isDone is not set');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
     'priority' => 'normal'
@@ -52,7 +52,7 @@ $guy->seeResponseContainsJson(['id' => 6]);
 $guy->sendGET('/tasks/6');
 $guy->seeResponseContainsJson([
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 500
 ]);
 
@@ -60,7 +60,7 @@ $guy->seeResponseContainsJson([
 $guy->wantTo('priority is `low`');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 'low'
 ]);
 $guy->seeResponseIsJson();
@@ -69,7 +69,7 @@ $guy->seeResponseContainsJson(['id' => 7]);
 $guy->sendGET('/tasks/7');
 $guy->seeResponseContainsJson([
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 100
 ]);
 
@@ -77,7 +77,7 @@ $guy->seeResponseContainsJson([
 $guy->wantTo('`priority` is `high`');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 'high'
 ]);
 $guy->seeResponseIsJson();
@@ -86,7 +86,7 @@ $guy->seeResponseContainsJson(['id' => 8]);
 $guy->sendGET('/tasks/8');
 $guy->seeResponseContainsJson([
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 900
 ]);
 
@@ -94,7 +94,7 @@ $guy->seeResponseContainsJson([
 $guy->wantTo('`priority` is not set');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
-    'marked'   => false
+    'isDone'   => false
 ]);
 $guy->seeResponseIsJson();
 $guy->seeResponseCodeIs(201);
@@ -102,7 +102,7 @@ $guy->seeResponseContainsJson(['id' => 9]);
 $guy->sendGET('/tasks/9');
 $guy->seeResponseContainsJson([
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 500
 ]);
 
@@ -110,7 +110,7 @@ $guy->seeResponseContainsJson([
 $guy->wantTo('card does not exist');
 $guy->sendPOST('/cards/900/tasks', [
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 'normal'
 ]);
 $guy->seeResponseIsJson();
@@ -119,7 +119,7 @@ $guy->seeResponseCodeIs(404);
 
 $guy->wantTo('name is not set');
 $guy->sendPOST('/cards/1/tasks', [
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 'normal'
 ]);
 $guy->seeResponseIsJson();
@@ -130,7 +130,7 @@ $guy->seeResponseJsonMatchesJsonPath('$.message');
 $guy->wantTo('`priority` is invalid');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
-    'marked'   => false,
+    'isDone'   => false,
     'priority' => 'bla'
 ]);
 $guy->seeResponseIsJson();
@@ -138,10 +138,10 @@ $guy->seeResponseCodeIs(400);
 $guy->seeResponseJsonMatchesJsonPath('$.message');
 
 
-$guy->wantTo('`marked` is invalid');
+$guy->wantTo('`isDone` is invalid');
 $guy->sendPOST('/cards/1/tasks', [
     'name'     => 'something',
-    'marked'   => 'asdas',
+    'isDone'   => 'asdas',
     'priority' => 'normal'
 ]);
 $guy->seeResponseCodeIs(201);
@@ -150,6 +150,6 @@ $guy->seeResponseContainsJson(['id' => 10]);
 $guy->sendGET('/tasks/10');
 $guy->seeResponseContainsJson([
     'name'     => 'something',
-    'marked'   => true,
+    'isDone'   => true,
     'priority' => 500
 ]);